demo2/requirements.txt
Django==5.1.2
django-allauth==65.3.1
django-allauth[socialaccount]==65.3.1
